# KYC360_InternshipAssessment/Models/Address.cs
namespace KYC360_InternshipAssessment.Models
{
    public class Address
    {
        public string? AddressLine { get; set; }
        public string? City { get; set; }
        public string? Country { get; set; }
    }
}