# KYC360_InternshipAssessment/Models/Name.cs
namespace KYC360_InternshipAssessment.Models
{
    public class Name
    {
        public string? FirstName { get; set; }
        public string? MiddleName { get; set; }
        public string? Surname { get; set; }

    }
}